pyproject.toml
[project]
name = "lily-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.3",
    "dotenv>=0.9.9",
    "openai>=1.102.0",
    "pillow>=11.3.0",
]
